UG.PyServices/vk_scraper/requirements.txt
vk-api==11.8.0
PyYAML==5.3.1
mysql-connector-python==8.0.20
